Allow to specify a proxy or use the env
Somehow OAuth1 does not automatically use the ENV to
use the “default” proxy. So use the default proxy
settings or use the explicitly specified one. 

This code does make it virtually impossible to not 
use a proxy if there is one defined in the environment
variables, but that would be a bit weird, right?